requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
woocommerce==3.0.0
odoo-rpc-client==0.9.0
pydantic==2.5.2
psycopg2-binary==2.9.9 